# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CommuteAds Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /devices:
    get:
      operationId: listDevices
      tags: [devices]
      summary: List all fleet devices
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [ACTIVE, STANDBY, OFFLINE]
      responses:
        "200":
          description: List of devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
    post:
      operationId: createDevice
      tags: [devices]
      summary: Register a new device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceInput"
      responses:
        "201":
          description: Created device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/{id}:
    get:
      operationId: getDevice
      tags: [devices]
      summary: Get a device by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          description: Not found
    patch:
      operationId: updateDevice
      tags: [devices]
      summary: Update device metadata
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceUpdate"
      responses:
        "200":
          description: Updated device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"

  # ── Telemetry ──────────────────────────────────────────────────────────────
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all campaigns
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [ACTIVE, PENDING, COMPLETED]
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a new campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}/approve:
    post:
      operationId: approveCampaign
      tags: [campaigns]
      summary: Approve a pending campaign
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"

  # ── Ad Assets ──────────────────────────────────────────────────────────────
  /ad-assets:
    get:
      operationId: listAdAssets
      tags: [ad-assets]
      summary: List ad assets
      parameters:
        - in: query
          name: campaign_id
          schema:
            type: integer
      responses:
        "200":
          description: List of ad assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdAsset"
    post:
      operationId: createAdAsset
      tags: [ad-assets]
      summary: Upload a new ad asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdAssetInput"
      responses:
        "201":
          description: Created ad asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdAsset"
  /ad-assets/{id}:
    get:
      operationId: getAdAsset
      tags: [ad-assets]
      summary: Get an ad asset by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ad asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdAsset"
        "404":
          description: Not found
  /ad-assets/{id}/approve:
    post:
      operationId: approveAdAsset
      tags: [ad-assets]
      summary: Approve an ad asset
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved ad asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdAsset"

  # ── Dynamic Content ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Device:
      type: object
      required:
        - id
        - nodeCode
        - batteryPct
        - isCharging
        - chargingModeStatus
        - signalDbm
        - currentLat
        - currentLng
        - screenStatus
        - lastPing
        - firmwareVersion
      properties:
        id:
          type: integer
        nodeCode:
          type: string
        batteryPct:
          type: integer
        isCharging:
          type: boolean
        chargingModeStatus:
          type: string
          enum: [PAUSED_AT_80, CHARGING_BELOW_50, NORMAL]
        signalDbm:
          type: integer
        currentLat:
          type: number
        currentLng:
          type: number
        currentlyPlaying:
          type: ["string", "null"]
        screenStatus:
          type: string
          enum: [ACTIVE, STANDBY]
        lastPing:
          type: string
        firmwareVersion:
          type: string
        hasSosAlert:
          type: ["boolean", "null"]
    DeviceInput:
      type: object
      required: [nodeCode, firmwareVersion]
      properties:
        nodeCode:
          type: string
        firmwareVersion:
          type: string
        currentLat:
          type: number
        currentLng:
          type: number
    DeviceUpdate:
      type: object
      properties:
        screenStatus:
          type: string
          enum: [ACTIVE, STANDBY]
        firmwareVersion:
          type: string

    # ── Telemetry ─────────────────────────────────────────────────────────────
...
    Campaign:
      type: object
      required:
        - id
        - advertiserId
        - name
        - targetFleetTier
        - totalTargetImpressions
        - deliveredImpressions
        - startDate
        - endDate
        - dailyCap
        - status
      properties:
        id:
          type: integer
        advertiserId:
          type: integer
        name:
          type: string
        targetFleetTier:
          type: string
        totalTargetImpressions:
          type: integer
        deliveredImpressions:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        dailyCap:
          type: integer
        status:
          type: string
          enum: [ACTIVE, PENDING, COMPLETED]
        pacingPercent:
          type: ["number", "null"]
    CampaignInput:
      type: object
      required:
        - advertiserId
        - name
        - targetFleetTier
        - totalTargetImpressions
        - startDate
        - endDate
        - dailyCap
      properties:
        advertiserId:
          type: integer
        name:
          type: string
        targetFleetTier:
          type: string
        totalTargetImpressions:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        dailyCap:
          type: integer
    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
        targetFleetTier:
          type: string
        totalTargetImpressions:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        dailyCap:
          type: integer
        status:
          type: string
          enum: [ACTIVE, PENDING, COMPLETED]

    # ── Ad Asset ──────────────────────────────────────────────────────────────
    AdAsset:
      type: object
      required: [id, campaignId, title, videoUrl, durationSec, isApproved]
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        title:
          type: string
        videoUrl:
          type: string
        durationSec:
          type: integer
        isApproved:
          type: boolean
    AdAssetInput:
      type: object
      required: [campaignId, title, videoUrl, durationSec]
      properties:
        campaignId:
          type: integer
        title:
          type: string
        videoUrl:
          type: string
        durationSec:
          type: integer

    # ── Dynamic Content ───────────────────────────────────────────────────────
...