# lib/api-zod/src/generated/types/adAssetInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CommuteAds Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AdAssetInput {
  campaignId: number;
  title: string;
  videoUrl: string;
  durationSec: number;
}